started reimplementing openfin
most of it is mostly working. I need to find out how to emulate history.push withing the openfin enviroment. Clicking on a peer from a popout window does not work. Poping out an element defaults to apple to start.

The work is sortof hacky right now, my plan is to get it working, then make it better,